ai-agent-python/requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
Pillow>=9.0.0